make NamespacesCount int64 due to consistency with other values and in case of int overflow

+++ internal/databases/mongo/binary/metadata_collector.go
@@ -15,7 +15,7 @@ type StorageMetadataCollector struct {
 	mongodService *MongodService
 	pmc           *PartialMetadataCollector
 	heap          *NsSizeHeap
-	counter       int
+	counter       int64
 	top100Ns      *[]string
 	TarsChan      chan internal.TarFileSets
 	ErrsChan      chan error
